refactor: extract listing generator hook and ui components
