+++ app/components/index.js
@@ -1,3 +1,10 @@
-export { OptionPicker } from "./option-picker";
-export { PromptPreviewCard } from "./prompt-preview-card";
+export { AdminReviewPanel } from "./admin-review-panel";
+export { AdminToolsCard } from "./admin-tools-card";
+export { DescriptionSettings } from "./description-settings";
+export { GarmentTypeSelector } from "./garment-type-selector";
 export { InfoTooltip } from "./info-tooltip";
+export { OptionPicker } from "./option-picker";
+export { PoseStatusList } from "./pose-status-list";
+export { PromptPreviewCard } from "./prompt-preview-card";
+export { SceneModelOptions } from "./scene-model-options";
+export { UploadPanel } from "./upload-panel";
